Remove empty parser files.
