--- readmeai/parsers/cicd/bitbucket.py
@@ -1,1 +1,0 @@
-"""Parser for Bitbucket Pipelines (bitbucket-pipelines.yml) configuration files."""
