--- readmeai/parsers/cicd/circleci.py
@@ -1,1 +1,0 @@
-"""Parser for .circleci/config.yml configuration files."""
