--- readmeai/parsers/cicd/github.py
@@ -1,1 +1,0 @@
-"""Parser for GitHub Actions (.github/workflows/) configuration files."""
