--- readmeai/parsers/cicd/gitlab.py
@@ -1,1 +1,0 @@
-"""Parser for .gitlab-ci.yml configuration files."""
