--- readmeai/parsers/cicd/jenkins.py
@@ -1,1 +1,0 @@
-"""Parser for Jenkinsfile (Jenkinsfile) configuration files."""
