--- readmeai/parsers/cicd/travis.py
@@ -1,1 +1,0 @@
-"""Parser for .travis.yml configuration files."""
