--- readmeai/parsers/configuration/ansible.py
@@ -1,1 +1,0 @@
-"""Parser for Ansible (playbook.yml, ansible/site.yml) configuration files."""
