--- readmeai/parsers/configuration/apache.py
@@ -1,1 +1,0 @@
-"""Parser for apache.py (httpd.conf) configuration files."""
