--- readmeai/parsers/configuration/nginx.py
@@ -1,1 +1,0 @@
-"""Parser for Nginx (nginx.conf) configuration files."""
