--- readmeai/parsers/infrastructure/cloudformation.py
@@ -1,1 +1,0 @@
-"""Parser for cloudformation.yaml (AWS CloudFormation) configuration files."""
