--- readmeai/parsers/infrastructure/terraform.py
@@ -1,1 +1,0 @@
-"""Parser for main.tf (Terraform) configuration files."""
