--- readmeai/parsers/orchestration/kubernetes.py
@@ -1,1 +1,0 @@
-"""Parser for Kubernetes (k8s.yml) configuration files."""
