--- readmeai/parsers/package/composer.py
@@ -1,1 +1,0 @@
-"""Parser for PHP Composer (composer.json) configuration files."""
