--- readmeai/parsers/package/gem.py
@@ -1,1 +1,0 @@
-"""Parser for Gemfile.lock (Ruby) configuration files."""
