--- readmeai/parsers/package/nuget.py
@@ -1,1 +1,0 @@
-"""Parser for NuGet.Config (.NET) configuration files."""
