--- readmeai/parsers/package/pip.py
@@ -1,1 +1,0 @@
-"""Parser for Pip (requirements.txt, Pipfile) configuration files."""
