--- tests/parsers/package/test_yarn.py
@@ -1,1 +1,0 @@
-"""Tests for the yarn.lock file parser module."""
